serial onb interface fix

+++ onb/serialonbinterface.h
@@ -26,7 +26,7 @@ public:
 
     int availableWriteCount();
 
-    int addFilter(uint32_t id, uint32_t mask=0xFFFFFFFF);
+    int addFilter(unsigned long id, unsigned long mask=0xFFFFFFFF);
     void removeFilter(int number);
 
 signals:
